+++ src/App.js
@@ -1,12 +1,12 @@
 import React, { Component } from 'react';
 import { BrowserRouter as Router, Route, Link } from 'react-router-dom';
 
-import Header from './components/Header';
-import Search from './components/Search';
-import Events from './components/Events';
-import Details from './components/Details';
-import Map from './components/Map';
+import Search from './components/Search/Search';
+import Events from './components/Events/Events';
+import Details from './components/Details/Details';
+import Map from './components/Map/Map';
 import './App.scss';
+import Header from './components/Header/Header';
 
 class App extends Component {
   render() {
