--- src/components/Action.js
@@ -1,13 +1,0 @@
-import React, {Component} from 'react';
-
-
-class Action extends Component {
-  render() {
-    return (
-      <div className="Action">
-      </div>
-    );
-  }
- }
-
-export default Action;
