+++ src/components/Action/Action.js
@@ -1,0 +1,9 @@
+import React, { Component } from 'react';
+
+class Action extends Component {
+  render() {
+    return <div className="Action" />;
+  }
+}
+
+export default Action;
