--- src/components/Data.js
@@ -1,38 +1,0 @@
-export default [
-  {
-    id: 1234455,
-    title: "The State of Spanish and European VC with Tech.eu and Novobrief",
-    date: "27/02/2018",
-    description: "Tech and Novobrief will share our new full-year data on venture capitall activity across Europe, with a focus on Spanin and our outlook for 2017",
-    address: "c/ Moreno Nieto 2, 28005, Madrid", 
-    prices: "10 €",
-    time: "18:00h"
-  },
-  {
-    id: 2,
-    title: "Data Sciencie",
-    date: "27/02/2018",
-    description: "Tech and Novobrief will share our new full-year data on venture capitall activity across Europe, with a focus on Spanin and our outlook for 2017",
-    address: "c/ Moreno Nieto 2, 28005, Madrid",
-    prices: "10 €",
-    time: "18:00h"
-  },
-  {
-    id: 3,
-    title: "The Stateee of Spanish and European VC with Tech.eu and Novobrief",
-    date: "27/02/2018",
-    description: "Tech and Novobrief will share our new full-year data on venture capitall activity across Europe, with a focus on Spanin and our outlook for 2017",
-    address: "c/ Moreno Nieto 2, 28005, Madrid",
-    prices: "10 €",
-    time: "18:00h"
-  },
-  {
-    id: 4,
-    title: "Amazing React",
-    date: "27/02/2018",
-    description: "Tech and Novobrief will share our new full-year data on venture capitall activity across Europe, with a focus on Spanin and our outlook for 2017",
-    address: "c/ Moreno Nieto 2, 28005, Madrid",
-    prices: "10 €",
-    time: "18:00h"
-  }
-];
