--- src/components/Detail.js
@@ -1,12 +1,0 @@
-import React, {Component} from 'react';
-
-class Detail extends Component {
-  render() {
-    return (
-      <div className="Detail">
-      </div>
-    );
-  }
- }
-
-export default Detail;
