+++ src/components/Detail/Detail.js
@@ -1,0 +1,9 @@
+import React, { Component } from 'react';
+
+class Detail extends Component {
+  render() {
+    return <div className="Detail" />;
+  }
+}
+
+export default Detail;
