+++ src/components/DetailButton/DetailButton.scss
