+++ src/components/Event/Event.scss
