+++ src/components/Map/Map.js
@@ -1,15 +1,14 @@
-import React, {Component} from 'react';
-import Marker from './Marker';
-
+import React, { Component } from 'react';
+import Marker from '../Marker/Marker';
 
 class Map extends Component {
   render() {
     return (
       <div className="Map">
-        <Marker/>
+        <Marker />
       </div>
     );
   }
- }
+}
 
 export default Map;
