--- src/components/Marker.js
@@ -1,14 +1,0 @@
-import React, {Component} from 'react';
-import Popup from './Popup';
-
-
-class Marker extends Component {
-  render() {
-    return (
-      <div className="Marker">
-      </div>
-    );
-  }
- }
-
-export default Marker;
