+++ src/components/Marker/Marker.js
@@ -1,0 +1,10 @@
+import React, { Component } from 'react';
+import Popup from '../Popup/Popup';
+
+class Marker extends Component {
+  render() {
+    return <div className="Marker" />;
+  }
+}
+
+export default Marker;
