--- src/components/Popup.js
@@ -1,13 +1,0 @@
-import React, {Component} from 'react';
-
-
-class Popup extends Component {
-  render() {
-    return (
-      <div className="Popup">
-      </div>
-    );
-  }
- }
-
-export default Popup;
