+++ src/components/Popup/Popup.js
@@ -1,0 +1,9 @@
+import React, { Component } from 'react';
+
+class Popup extends Component {
+  render() {
+    return <div className="Popup" />;
+  }
+}
+
+export default Popup;
