--- src/components/Search.js
@@ -1,13 +1,0 @@
-import React, {Component} from 'react';
-
-
-class Search extends Component {
-  render() {
-    return (
-      <div className="Search">
-      </div>
-    );
-  }
- }
-
-export default Search;
