+++ src/components/Search/Search.js
@@ -1,0 +1,9 @@
+import React, { Component } from 'react';
+
+class Search extends Component {
+  render() {
+    return <div className="Search" />;
+  }
+}
+
+export default Search;
