+++ src/components/Shared/Time/Time.js
@@ -1,0 +1,9 @@
+import React, { Component } from 'react';
+
+class Time extends Component {
+  render() {
+    return <div className="Time" />;
+  }
+}
+
+export default Time;
