--- src/components/Social.js
@@ -1,13 +1,0 @@
-import React, {Component} from 'react';
-
-
-class Social extends Component {
-  render() {
-    return (
-      <div className="Social">
-      </div>
-    );
-  }
- }
-
-export default Social;
