+++ src/components/Social/Social.js
@@ -1,0 +1,9 @@
+import React, { Component } from 'react';
+
+class Social extends Component {
+  render() {
+    return <div className="Social" />;
+  }
+}
+
+export default Social;
