--- src/components/Time.js
@@ -1,13 +1,0 @@
-import React, {Component} from 'react';
-
-
-class Time extends Component {
-  render() {
-    return (
-      <div className="Time">
-      </div>
-    );
-  }
- }
-
-export default Time;
